Add Windows-friendly alternative for our checkup status emojis
