+++ ee/debug/checkups/checkups_other.go
@@ -1,0 +1,21 @@
+//go:build !windows
+// +build !windows
+
+package checkups
+
+func (s Status) Emoji() string {
+	switch s {
+	case Informational:
+		return " "
+	case Passing:
+		return "✅"
+	case Warning:
+		return "⚠️"
+	case Failing:
+		return "❌"
+	case Erroring:
+		return "❌"
+	default:
+		return "? "
+	}
+}
